rag/requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
prometheus-fastapi-instrumentator
